feat: add workspace hub controller and shell controller hooks
- Implemented `useWorkspaceHubController` to manage hub templates and errors.
- Created `useWorkspaceShellController` for handling workspace shell state and interactions.
- Refactored `useWorkspaceNavigation` to remove unnecessary state management for channel tools and member list.
- Updated tests to reflect changes in navigation hook usage.
- Adjusted legacy contract test to ensure correct hub template creation behavior.
