+++ web/app/src/hooks/workspace/index.ts
@@ -1,9 +1,15 @@
 export * from "./WorkspaceControllerContext";
 export * from "./useCLIProxyAuthStatuses";
+export * from "./useAgentController";
+export * from "./useConversationController";
+export * from "./useProfilePreviewController";
 export * from "./useProfileModelOptions";
+export * from "./useUpgradeController";
 export * from "./useWorkspaceController";
 export * from "./useWorkspaceData";
+export * from "./useWorkspaceHubController";
 export * from "./useWorkspaceHubSelection";
 export * from "./useWorkspaceNavigation";
+export * from "./useWorkspaceShellController";
 export * from "./workspaceQueries";
 export * from "./workspaceUiStore";
